ВЧЕРА fixed... ВСЕ fixed
